Refactor to have intermediate function when converting molecule formats
In order to accomodate more formats in the future, a function is used to
first convert from, for example, smiles to mol, then the mol can be
serialized.
The intermediate functions could also be mol block to mol, etc.

In the opposite direction, a molecule can be deserialized back to the mol. From
there, it can be possible to have functions convert from mol to a final
format, such as mol_to_smiles, or mol_to_mol_block,
etc.

mol_block/smiles/etc -> mol -> binary mol -> mol_block/smiles/etc