working version of dalke fp only being generated when needed for substructure search and caching it for the query
The dalke fp takes a bit of time to generate. For queries that come in
that are not already UmbraMols, like umbra_is_substruct(umbra_mol,
'CC'), the dalke fp was being generated multiple times in a query, and
this takes a lot of time. Caching this in a std::map seems to work.

Also, the dalke fp generation is totally unneccessary for the exact
match. This commit also has a working version that moves the dalke fp
generation code out of the constructor for UmbraMol. This allows for the
dalke fp generation method to be called only when doing substructure
matches, and otherwise, no dalkefp will be generated when not needed.
This seems to save a lot of time, and return the umbra_is_exact_match
performance back to the original with the counts