set validity mask and display warning when mol cannot be created from smiles
Previously, when inserting many records into a Mol column, if a cast or
mol_from_smiles fails, the entire insert would fail. This might be
frustrating if inserting large amounts and then having it fail at the
end and then requiring the user to figure out which SMILES fails.

Instead, now the validity mask is used to mark that entry as invalid,
and that will insert that particular record as null. Then the insert
would not fail if there are some invalid records. These invalid records
could be cleaned up later by the user if desired.

A warning is also displayed so that the user can see which SMILES failed