initial sdf replacement scan
The replacement scan allows such queries:
`SELECT * FROM 'test.sdf';`
or `SELECT mol FROM 'test.sdf';`

by detecting the sdf file extension and running the appropriate code.

Specifying the columns is not necessary, but the binder will need to auto
detect the columns since the user does not specify
them via a parameter.

Currently, the column detection just assumes that all records in the
file have the same properties. It will read the first record and then
use the property list there as the column names.

Projection pushdown is not yet implemented