Remove count prefix and modify substructure search bail out
The count prefix is removed and the prefix to the umbra mol now contains
only the dalke fp for substructure screens. Previously, the dalke fp was
after the count prefix and could only be accessed after a pointer chase.
Now the first 4 bytes of the dalke fp is inlined and can be screened
already without the pointer chase